requirements.txt
python-telegram-bot==21.5
anthropic>=0.34.0
python-dotenv==1.0.1
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
matplotlib>=3.8.0
Pillow>=10.0.0
notion-client==2.2.1
